Rename _math to utils

+++ utils.py
@@ -1,7 +1,6 @@
 import os
 import random
 import uuid
-from decimal import Decimal
 
 from PIL import Image
 import conversor
@@ -29,4 +28,4 @@ def estimate_size(width: int, height: int, path: str, bw: bool) -> str:
         sizes.append(os.stat(_path).st_size)
         os.remove(_path)
 
-    return conversor.to_readable(Decimal((sum(sizes) / 3) * ammount_of_images(width * height, bw)))
+    return conversor.to_readable((sum(sizes) / 3) * ammount_of_images(width * height, bw))
